# swagger.yaml
swagger: '2.0'
info:
  version: '1.0'
  title: "HUBusivi"
  description: REST Api to manage the exams
  license:
    name: MIT
    url: https://github.com/apiaryio/polls-api/blob/master/LICENSE
host: ingegneriasoftwa2.herokuapp.com
basePath: /api/v2
schemes:
- https
consumes:
- application/json
produces:
- application/json 

paths:
  /students:
    post:
      tags: 
      - "Students"
      description: Adds a student on the database
      operationId: addStudent
      parameters:
        - name: id
          in: query
          description: id number of student
          type: string
          x-exaple: "128888"
        - name: email
          in: query
          description: data of the student
          type: string
          x-example: "pincopallino@gmail.com"
        - name: name
          in: query
          description: name of student
          type: string
          x-example: "Pinco"
        - name: surname
          in: query
          description: surname of student
          type: string
          x-example: "Pallino"
        - name: password
          in: query
          description: password for the login
          type: string
          x-example: "ppallino"
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Students'
    
  
  /students/{Id}:
    parameters:
        - name: Id
          in: path
          description: ID of the student to fetch
          required: true
          type: string
          x-example: '1234'
    get:
      description: Returns a student from the database by ID
      operationId: getStudent
      
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful Response
          schema:
            type: array
            items:
              $ref: '#/definitions/Students'
     
    delete:
      description: Removes a student from the database
      operationId: removeStudent
      parameters:
        - name: id_student
          in: query
          description: id number of student that I want to delete
          type: integer
          x-example: 1
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
  
  /teacher:
    post:
      tags:
      - "Teacher"
      description: Adds a teacher
      operationId: Add Assignment
      parameters:
        - name: id
          in: query
          description: id of the teacher
          type: integer
          x-exaple: 128888
        - name: email
          in: query
          description: email of the teacher
          type: string
          x-example: "prof@gmail.com"
        - name: name
          in: query
          description: name of teacher
          type: string
          x-example: "Giorgio"
        - name: surname
          in: query
          description: surname of the teacher
          type: string
          x-example: "Rodi"
        - name: password
          in: query
          description: password for the login
          type: string
          x-example: "grodi"
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
    get:
      description: Returns a teacher from the database by ID
      operationId: getTeacher
      parameters:
        - name: id
          in: query
          description: id of the teacher
          type: integer
          x-example: 198888
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
          schema:
            type: array
            items:
              $ref: '#/definitions/Teachers'
    
  /teacher/{Id}:
    parameters:
      - name: Id
        in: path
        description: ID of teacher to fetch
        required: true
        type: string
        x-example: '1234'
    
    get:
        description: Returns a teacher from the database by ID
        operationId: getSpecTeacher
        responses:
          400:
            description: Bad request
          404:
            description: Resources not found
          200:
            description: Successful response
            schema:
              type: array
              items:
                $ref: '#/definitions/Teachers'
    
    delete:
      description: Removes a teacher from the database
      operationId: RemoveTeacher
      parameters:
        - name: id_teacher
          in: query
          description: ID of teacher that I want to delete
          type: integer
          x-example: 1
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
  
  /assignment:
    post: 
        tags: 
        - "Assignment"
        description: Adds an Assignment(only teacher)
        operationId: addAssignment
        parameters:
          - name: Id
            in: query
            description: id of the assignment
            type: string
            x-example: "123"
          - name: name
            in: query
            description: name of the assignment
            type: string
            x-example: "Assignment1" 
          - name: description
            in: query
            description: description of the assignment
            type: string
            x-example: "description"
          - name: upload date
            in: query
            description: upload date of the assignment
            type: string
            x-example: "12/12/2018"
          - name: expire date
            in: query
            description: Expire date of the assignment
            type: string
            x-example: "12/12/2018"
          - name: id_course
            in: query
            description: id of the course
            type: string
            x-example: "186495"
          
        responses:
          400:
            description: Bad request
          404:
            description: Resources not found
          200:
            description: Successful response
            

  /assignment/{Id}:  
    parameters:
      - name: Id
        in: path
        description: ID of the assignment to fetch
        required: true
        type: string
        x-example: '1234'
    put:
      description: Updates an assignment
      operationId: updateAssignment
      parameters:
          - name: Delivery date
            in: query
            description: Delivery date of assignment
            type: string
            x-example: "12/12/2018"
          - name: Expire date
            in: query
            description: Expire date of assignment
            type: string
            x-example: "12/12/2018"
          - name: courseId
            in: query
            description: id of the course
            type: string
            x-example: "186495"
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
          
    delete:
      description: Removes an assignment
      operationId: removeAssignment
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response  
    
  /assignment/avg:
    post:
      description: >-
        Calculates the marks average about a specific assignment of a course(only teacher).
      operationId: avgMark
      parameters:
        - name: Id
          in: query
          description: Course which I want to get the average of marks
          type: integer
          x-example: 113355
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Marks average
          
  
  /marks:
    post:
      tags:
      - "Marks"
      description: Adds a Mark (only teacher)
      operationId: addMark
      parameters:
        - name: mark
          in: query
          description: mark of the assignment
          type: integer
          x-example: 22
        - name: date
          in: query
          description: date of the given mark
          type: string
          x-example: "05/12/2018"
        - name: id_student
          in: query
          description: id of the student 
          type: integer
          x-example: 123456
        - name: id_mark
          in: query
          description: mark id of the mark
          type: integer
          x-example: 113355
        - name: id_assignment
          in: query
          description: assignment id of the mark
          type: integer
          x-example: 113325
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
    
    put:
      description: Updates a mark of a specific exam of a student (only teacher)
      operationId: updateMark
      parameters:
        - name: mark
          in: query
          description: mark to update
          type: integer
          x-example: 22
        - name: id_student
          in: query
          description: id of the student
          type: integer
          x-example: 123456
        - name: id_course
          in: query
          description: course id of the mark
          type: integer
          x-example: 113355
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
    
    delete:
      description: Removes a Mark of a specific exam of a student (only teacher)
      operationId: removeMark
      parameters:
        - name: id_student
          in: query
          description: matricola of the student
          type: integer
          x-example: 123456
        - name: id_course
          in: query
          description: course id of the mark
          type: integer
          x-example: 113355
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response

    
  
  /marks/courseId:
    get:
      description: Gets your marks of a specific course
      operationId: getMark
      parameters:
        - name: id_student
          in: query
          description: id of the student
          type: integer
          x-example: 123456
        - name: id_course
          in: query
          description: course id of the mark
          type: integer
          x-example: 113355
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
  
  /courses:
    put:
      description: Adds a course(teacher)/Join a course(student)
      operationId: addCourse
      parameters:
        - name: Id
          in: query
          description: course data
          type: string
          x-example: 'Ingegneria del software 2'
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
  
    get:
      tags:
      - "Courses"
      description: Gets the list of all your courses
      operationId: coursesList
      parameters:
        - name: Id
          in: query
          description: id of the student
          type: integer
          x-example: 123456
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: List of all your courses
          schema:
              type: array
              items:
                $ref: '#/definitions/Courses'


  /courses/{Id}:  
    parameters:
      - name: Id
        in: path
        description: ID of courses to fetch
        required: true
        type: string
        x-example: '1234'
    put:
      description: Updates a Course(teacher only)
      operationId: updateCourse
      parameters:
        - name: course name
          in: query
          description: course name
          type: string
          x-example: 'Ingegneria del software - 2'
        - name: id_course
          in: query
          description: course id
          type: integer
          x-example: 113355
        - name: description
          in: query
          description: course description
          type: string
          x-example: "description"
        - name: start
          in: query
          description: start of the course
          type: string
          x-example: "22/10/2018"
        - name: id_teacher
          in: query
          description: id teacher of the course
          type: integer
          x-example: 163355
        - name: end
          in: query
          description: end of the course
          type: string
          x-example: "23/04/2019"
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
          
    delete:
      description: Removes a course(teacher only)/Leave a course(student only)
      operationId: removeCourse
      
      responses:
        400:
          description: Bad request
        404:
          description: Resources not found
        200:
          description: Successful response
  
    

definitions:
  Students:
    type: object
    properties:
      id:
        type: string
      email:
        type: string
      name:
        type: string
      surname:
        type: string
      password:
        type: string
    required:
      - id
      - email
      - name
      - surname
      - password
    example:
      id: 12345
      email: stefano@gmail.com
      name: Stefano
      surname: Pretto
      password: stefano
  Teachers:
    type: object
    properties:
      id:
        type: string
      email:
        type: string
      name:
        type: string
      password:
        type: string
    required:
      - id
      - email
      - name
      - password
    example:
      id: 12345
      email: marco@gmail.com
      name: marco
      password: marco
  Assignment:
    type: object
    properties:
      id:
        type: string
      text:
        type: string
      name:
        type: string
      expireData:
        type: string
      uploadData:
        type: string
      studentId:
        type: string
      courseId:
        type: string
    required:
      - id
      - text
      - name
      - expireData
      - uploadData
      - studentId
      - courseId
    example:
      id: 12345
      text: Risolvi l'integrale
      name: Analisi
      expireData: 13-12-2018
      uploadData: 11-12-2018
      studentId: 123
      courseId: 1234
  Courses:
    type: object
    properties:
      s_id:
        type: string
      description:
        type: string
      name:
        type: string
      t_id:
        type: string
    required:
      - s_id
      - description
      - name
      - t_id
    example:
      s_id: 12345
      description: Informatica+Economia
      name: INF ORG
      t_id: 123456